backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.108.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
pytz>=2023.3
python-dateutil>=2.8.0

# Core LLM dependencies only
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-groq>=0.3.5
langchain-google-genai>=0.0.5
openai>=1.10.0
groq>=0.4.0
google-generativeai>=0.3.0
anthropic>=0.7.0 